Update keyer_pin_settings_opencwkeyer_mk2.h
The Open CW Keyer MK2 does not have any circuitry to support a second transmitter. Clearing the pin defines for this non-existent transmitter removes the possibility of the user accidentally having tx2 enabled. A long press of the memory 2 button will set the keyer up for tx2 output and there is no indication that this is the configuration. With no pins for tx2 then inadvertently selecting tx2 is not possible.